pioi

+++ oa/src/main/java/com/zhm/rabbit/oa/repositories/dao/OaMenuRepository.java
@@ -3,11 +3,10 @@ package com.zhm.rabbit.oa.repositories.dao;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
-import org.springframework.transaction.annotation.Transactional;
 
 import com.zhm.rabbit.oa.repositories.OaMenu;
 
-public interface OaMenuRepository extends  CrudRepository<OaMenu, Long>{
+public interface OaMenuRepository extends  CrudRepository<OaMenu, Integer>{
 
 	List<OaMenu> findByType(int type);
